feat(sign): delete placeholder & add trailingIcon
